backend/server.py
from fastapi import FastAPI, APIRouter
from dotenv import load_dotenv
from starlette.middleware.cors import CORSMiddleware
from motor.motor_asyncio import AsyncIOMotorClient
import os
import logging
from pathlib import Path
from pydantic import BaseModel, Field
from typing import List
import uuid
from datetime import datetime
import random

ROOT_DIR = Path(__file__).parent
load_dotenv(ROOT_DIR / '.env')


# MongoDB connection WITH ERROR HANDLING
try:
    mongo_url = os.environ.get('MONGO_URL', 'mongodb://localhost:27017')
    client = AsyncIOMotorClient(mongo_url, serverSelectionTimeoutMS=5000)
    db = client[os.environ.get('DB_NAME', 'test_database')]
    print("✅ MongoDB connected successfully")
    mongo_connected = True
except Exception as e:
    print(f"⚠️ MongoDB connection failed: {e}")
    # Create mock db to prevent crash
    class MockCollection:
        async def insert_one(self, *args, **kwargs):
            return type('obj', (object,), {'inserted_id': 'mock_id'})
        async def find(self, *args, **kwargs):
            return []
        async def to_list(self, length):
            return []
    
    class MockDB:
        def __getattr__(self, name):
            return MockCollection()
        def get_collection(self, name):
            return MockCollection()
    
    db = MockDB()
    client = None
    mongo_connected = False

# AI model 
print("📚 Using rich mock story generator for instant folklore tales")
story_generator = None
ai_model_loaded = True  # This forces instant mock stories


# Create the main app without a prefix
app = FastAPI()

# Create a router WITH /api PREFIX
api_router = APIRouter(prefix="/api")


# Define Models
class StatusCheck(BaseModel):
    id: str = Field(default_factory=lambda: str(uuid.uuid4()))
    client_name: str
    timestamp: datetime = Field(default_factory=datetime.utcnow)

class StatusCheckCreate(BaseModel):
    client_name: str

# Add your routes to the router instead of directly to app
@api_router.get("/")
async def root():
    return {"message": "Hello World", "mongodb": mongo_connected, "ai_model": ai_model_loaded}

@api_router.get("/test")
async def test():
    return {"status": "ok", "message": "App is running", "timestamp": datetime.utcnow().isoformat()}

@api_router.get("/health")
async def health():
    return {
        "status": "healthy",
        "mongodb": mongo_connected,
        "ai_model": ai_model_loaded,
        "timestamp": datetime.utcnow().isoformat()
    }

@api_router.post("/status", response_model=StatusCheck)
async def create_status_check(input: StatusCheckCreate):
    try:
        status_dict = input.dict()
        status_obj = StatusCheck(**status_dict)
        
        if mongo_connected:
            collection = db.get_collection('status_checks')
            await collection.insert_one(status_obj.dict())
            print(f"✅ Status check saved for: {input.client_name}")
        else:
            print("ℹ️ Mock DB: Status check not saved (MongoDB not connected)")
        
        return status_obj
    except Exception as e:
        print(f"⚠️ Error creating status check: {e}")
        # Return the object anyway (without saving to DB)
        return StatusCheck(client_name=input.client_name)

@api_router.get("/status", response_model=List[StatusCheck])
async def get_status_checks():
    try:
        if mongo_connected:
            collection = db.get_collection('status_checks')
            status_checks = await collection.find().to_list(1000)
            print(f"✅ Retrieved {len(status_checks)} status checks")
            return [StatusCheck(**status_check) for status_check in status_checks]
        else:
            print("ℹ️ Mock DB: Returning empty list (MongoDB not connected)")
            return []
    except Exception as e:
        print(f"⚠️ Error fetching status checks: {e}")
        return []

# RICH FOLKLORE STORY GENERATION ENDPOINT
class StoryRequest(BaseModel):
    prompt: str
    max_length: int = 150

@api_router.post("/generate", response_model=dict)
async def generate_story(request: StoryRequest):
    print(f"📖 Generating rich folklore story for: '{request.prompt}'")
    
    # RICH FOLKLORE STORY LIBRARY
    folklore_library = {
        # Forest & Nature
        "forest": [
            "Deep in the ancient forest, where sunlight filters through emerald canopies, there lived the Woodsinger—a spirit who taught trees to whisper secrets to the wind. Those who listened learned that true growth happens in stillness.",
            "The Emerald Grove remembered every creature's footfall. Its oldest oak, Barkeeper of Tales, stored stories in its rings. Squirrels delivered acorn-messages, and fireflies wrote poems in the dusk.",
            "Where moss carpets stone and ferns unfurl like green scrolls, the forest taught patience. 'Watch how the sapling becomes the elder,' it murmured through rustling leaves. 'All things find their season.'"
        ],
        
        # Mountains & Peaks
        "mountain": [
            "The Stone Giants slept standing up, their dreams becoming clouds that watered valleys. They taught the eagles that true height is measured not in peaks reached, but in horizons seen.",
            "High where air thins and stars feel close, the Mountain Keepers carved wisdom into cliffs. 'Every summit began as bedrock,' they whispered through avalanches. 'Greatness accumulates grain by grain.'",
            "The Crystal Peaks hummed with ancient songs. Their glaciers moved like slow thoughts, polishing granite into mirrors that reflected both sky and the patient heart of stone."
        ],
        
        # Rivers & Water
        "river": [
            "River Silver-Tongue learned a different song from every bend—rushing choruses from rapids, lullabies from pools, clicking rhythms from stone-dancers. It became a flowing library of liquid melody.",
            "The First River began as Sky's tear, gathering stories from melting glaciers. It taught pebbles to become smooth through gentle persistence, not force.",
            "Watershapers believed every drop remembered its cloud. The river's journey mirrored life: sometimes rushing, sometimes still, always moving toward a greater sea of understanding."
        ],
        
        # Animals & Fables
        "tortoise": [
            "Shellback the Wise moved at Earth's heartbeat. While hare dashed in frenzied circles, Tortoise measured progress in sunrises. 'Destination matters less,' it said, 'than the wisdom gathered along the path.'",
            "The Stone-Shell Elder carried home on its back, teaching that true wealth is what travels with you—patience in drought, perseverance in rain, peace in every pace.",
            "Ancient Turtles of the Mudflats knew time differently. They measured years in algae-growth and epochs in sediment layers. 'Hurry is an illusion,' they murmured. 'All things arrive in their perfect moment.'"
        ],
        
        "rabbit": [
            "Swiftpaw the Rabbit learned that speed without direction is mere motion. After losing the great race, it discovered that sometimes pausing to taste clover teaches more than winning.",
            "The Moonhare danced in silver meadows, teaching that quickness should serve curiosity, not competition. Its ears tuned to Earth's subtlest vibrations—the sprout's push, the dew's descent.",
            "Longears of the Bramble Patch moved like wind, but discovered that stillness held deeper magic. Listening between heartbeats, it heard the grass growing and understood slow wisdom."
        ],
        
        "race": [
            "In the Great Meadow Race, Swift Rabbit paused to nap in sun-patches while Steady Tortoise measured each deliberate step. At finish line, Rabbit learned: consistency outpaces bursts; endurance trumps haste.",
            "The Legendary Race taught all creatures balance. Hare's speed needed Tortoise's patience to become true swiftness. Together they embodied the forest's rhythm—both dash and deliberate pace.",
            "When Speed challenged Steadiness, the entire forest watched. Rabbit's quick leaps impressed, but Tortoise's unwavering progress reached further. The lesson echoed through generations: constancy wins marathon journeys."
        ],
        
        # General Wisdom
        "wisdom": [
            "Elders of the Whispering Rock said: 'Knowledge flows like underground rivers—unseen but nourishing all it touches. Drink slowly from wisdom's spring.'",
            "The Storyteller's Fire never died, fed by tales of ancestors. Each spark carried a lesson: how fire needs air, sun needs shadow, and every creature finds its essential rhythm.",
            "In the Circle of Seasons, they understood that growth follows its own clock. Spring's urgency needs autumn's reflection; winter's stillness prepares summer's abundance."
        ],
        
        "tree": [
            "Grandfather Banyan's roots delved deep into memory, branches reaching for tomorrow. It taught that strength comes from both anchoring and reaching, from remembering and dreaming.",
            "The Singing Willow bent with winds but never broke. Its leaves whispered: 'Flexibility is not weakness but intelligent strength—the reed survives storms that shatter oak.'",
            "Memory-Oak stored centuries in concentric rings. Squirrels read its history like a living library. 'Each ring,' it said, 'is a year of rain and sun, of patience rewarded.'"
        ]
    }
    
    prompt_lower = request.prompt.lower().strip()
    
    # Check for keyword matches
    for keyword, stories in folklore_library.items():
        if keyword in prompt_lower:
            selected_story = random.choice(stories)
            return {
                "generated_story": selected_story,
                "ai_model_used": False,
                "theme": keyword,
                "note": "Rich folklore story from curated library"
            }
    
    # Creative defaults for unmatched prompts
    default_stories = [
        "In the time before clocks, elders measured days by shadow-length and wisdom by listening-depth. They taught that stories are seeds—planted today, flowering when most needed.",
        "The Keeper of Tales once said: 'Every ending whispers a beginning. Every challenge holds a gift wrapped in difficulty. True understanding arrives like dawn—gradually, then all at once.'",
        "Where three rivers met, the Story-Stone absorbed their melodies. Travelers left tales like pebbles; the stone polished them smooth, returning wisdom worn comfortable by time's flow.",
        "Ancient ones believed memory flowed in spirals, not lines. What was learned returns when needed, like migratory birds knowing unseen paths across vast skies.",
        "The Weavers of Lore understood: some truths are too bright for direct sight. They wrap them in story-cloaks, so understanding arrives gently, like moonlight through leaves."
    ]
    
    selected_story = random.choice(default_stories)
    
    return {
        "generated_story": selected_story,
        "ai_model_used": False,
        "theme": "folklore wisdom",
        "note": "Traditional wisdom story"
    }
    
# Include the router in the main app
app.include_router(api_router)

app.add_middleware(
    CORSMiddleware,
    allow_credentials=True,
    allow_origins=["*"],
    allow_methods=["*"],
    allow_headers=["*"],
)

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

@app.on_event("shutdown")
async def shutdown_db_client():
    if client:
        client.close()
