backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
tzdata>=2024.2
motor>=3.3.1
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.2.0
python-multipart>=0.0.9
jq>=1.6.0
numpy>=1.26.0
pymongo>=4.5.0

# Google Gemini AI
google-genai>=1.0.0

# Legacy (optional - for fallback)
# --extra-index-url https://download.pytorch.org/whl/cpu
# torch>=2.0.0
# transformers>=4.36.0
